User smaller data type (`tinyint`) for `CodeSignature.Type` column
Migration updated

+++ SG.TestRunService/Migrations/20200604091226_AddCodeSignatureType.cs
@@ -7,7 +7,7 @@ namespace SG.TestRunService.Migrations
     {
         protected override void Up(MigrationBuilder migrationBuilder)
         {
-            migrationBuilder.AddColumn<int>(
+            migrationBuilder.AddColumn<byte>(
                 name: "Type",
                 table: "CodeSignature",
                 nullable: false,
